pyproject.toml
[tool.poetry]
name = "krzysztof-expert-agent"
version = "0.1.0"
description = ""
authors = ["Luke Skyward <luke.skyward.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.2"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
openai-agents = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
